fix: TeamHasPlayer PlayerOnTeam으로 이름변경 및 타자 순번 추가
