+++ BE/baseball-be/src/main/java/kr/codesquad/baseball/team/domain/Team.java
@@ -4,24 +4,23 @@ import kr.codesquad.baseball.player.domain.Player;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.MappedCollection;
 
-import java.util.HashSet;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 
 public class Team {
     @Id
     private Long id;
     private String name;
 
-    @MappedCollection(idColumn = "TEAM_ID")
-    private Set<TeamHasPlayer> players = new HashSet<>();
+    @MappedCollection(idColumn = "TEAM_ID", keyColumn = "HITTER_ORDER")
+    private List<PlayerOnTeam> players = new ArrayList<>();
 
     public Team(String name) {
         this.name = name;
     }
 
     public void addPlayers(List<Player> players) {
-        this.players.addAll(TeamHasPlayer.teamHasPlayers(players));
+        this.players.addAll(PlayerOnTeam.teamHasPlayers(players));
     }
 
     @Override
